all API routes protected

+++ src/app/api/members/[email]/route.ts
@@ -1,5 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { PrismaClient } from "@prisma/client";
+import { session } from "@/lib/auth";
 
 const prisma = new PrismaClient();
 
@@ -7,6 +8,11 @@ export async function DELETE(
   req: NextRequest,
   { params }: { params: { email: string } }
 ) {
+  const isSession = await session(req, {} as NextResponse);
+
+  if (!isSession) {
+    return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+  }
   await prisma.member.delete({
     where: { email: params.email },
   });
@@ -17,6 +23,11 @@ export async function PATCH(
   req: NextRequest,
   { params }: { params: { email: string } }
 ) {
+  const isSession = await session(req, {} as NextResponse);
+
+  if (!isSession) {
+    return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+  }
   const { name, email, roleId, discordid, githubid } = await req.json();
   const updatedMember = await prisma.member.update({
     where: { email: params.email },
